+++ sdk/azcore/runtime/policy_sas_credential_test.go
@@ -50,6 +50,25 @@ func TestSASCredentialPolicy_RequiresHTTPS(t *testing.T) {
 	require.Error(t, err)
 }
 
+func TestSASCredentialPolicy_AllowHTTP(t *testing.T) {
+	cred := exported.NewSASCredential("foo")
+
+	policy := NewSASCredentialPolicy(cred, "fake-auth", &SASCredentialPolicyOptions{
+		InsecureAllowCredentialWithHTTP: true,
+	})
+	require.NotNil(t, policy)
+
+	pl := exported.NewPipeline(shared.TransportFunc(func(req *http.Request) (*http.Response, error) {
+		return &http.Response{}, nil
+	}), policy)
+
+	req, err := NewRequest(context.Background(), http.MethodGet, "http://contoso.com")
+	require.NoError(t, err)
+
+	_, err = pl.Do(req)
+	require.NoError(t, err)
+}
+
 func TestSASCredentialPolicy_NilCredential(t *testing.T) {
 	const headerName = "fake-auth"
 	policy := NewSASCredentialPolicy(nil, headerName, nil)